Support arbitrary children in AccordionItem
